chore(www): Fix link hash for frontmatter section of page (#17123)
Heading of frontmatter section was changed to from "Including frontmatter" to "Frontmatter for metadata in markdown files" in df94c22ac01496e32ac1e61d055b2df95ab2d98b, but the hashed link in the second step was not updated.  This PR updates the link to match the new ID of the heading.